+++ tests/test_dates.py
@@ -29,25 +29,35 @@ def test_date_format_to_regex():
         assert pattern.match(date) is None
 
 
+def test_datetime_format_to_regex():
+    # Check on a Sentinel-1-like datetime format
+    date_format = "%Y%m%dT%H%M%S"
+    date = "20221204T005230"
+    pattern = _dates._date_format_to_regex(date_format)
+
+    # Test that the date matches the regular expression
+    assert pattern.match(date) is not None
+
+
 def test_get_dates():
     assert _dates.get_dates("20200303_20210101.int") == [
-        datetime.date(2020, 3, 3),
-        datetime.date(2021, 1, 1),
+        datetime.datetime(2020, 3, 3),
+        datetime.datetime(2021, 1, 1),
     ]
 
-    assert _dates.get_dates("20200303.slc")[0] == datetime.date(2020, 3, 3)
-    assert _dates.get_dates(Path("20200303.slc"))[0] == datetime.date(2020, 3, 3)
+    assert _dates.get_dates("20200303.slc")[0] == datetime.datetime(2020, 3, 3)
+    assert _dates.get_dates(Path("20200303.slc"))[0] == datetime.datetime(2020, 3, 3)
     # Check that it's the filename, not the path
-    assert _dates.get_dates(Path("/usr/19990101/asdf20200303.tif"))[0] == datetime.date(
-        2020, 3, 3
-    )
-    assert _dates.get_dates("/usr/19990101/asdf20200303.tif")[0] == datetime.date(
+    assert _dates.get_dates(Path("/usr/19990101/asdf20200303.tif"))[
+        0
+    ] == datetime.datetime(2020, 3, 3)
+    assert _dates.get_dates("/usr/19990101/asdf20200303.tif")[0] == datetime.datetime(
         2020, 3, 3
     )
 
     assert _dates.get_dates("/usr/19990101/20200303_20210101.int") == [
-        datetime.date(2020, 3, 3),
-        datetime.date(2021, 1, 1),
+        datetime.datetime(2020, 3, 3),
+        datetime.datetime(2021, 1, 1),
     ]
 
     assert _dates.get_dates("/usr/19990101/notadate.tif") == []
@@ -57,27 +67,48 @@ def test_get_dates_with_format():
     # try other date formats
     fmt = "%Y-%m-%d"
     assert _dates.get_dates("2020-03-03_2021-01-01.int", fmt) == [
-        datetime.date(2020, 3, 3),
-        datetime.date(2021, 1, 1),
+        datetime.datetime(2020, 3, 3),
+        datetime.datetime(2021, 1, 1),
+    ]
+
+    fmt = "%Y%m%dT%H%M%S"
+    # Check the OPERA name
+    fn = "OPERA_L2_CSLC-S1_T087-185678-IW2_20180210T232711Z_20230101T100506Z_S1A_VV_v1.0.h5"
+    assert _dates.get_dates(fn, fmt) == [
+        datetime.datetime(2018, 2, 10, 23, 27, 11),
+        datetime.datetime(2023, 1, 1, 10, 5, 6),
+    ]
+
+    # Check the Sentinel name
+    fn = "S1A_IW_SLC__1SDV_20221204T005230_20221204T005257_046175_05873C_3B80.zip"
+    assert _dates.get_dates(fn, fmt) == [
+        datetime.datetime(2022, 12, 4, 0, 52, 30),
+        datetime.datetime(2022, 12, 4, 0, 52, 57),
+    ]
+
+    # Check without a format using default
+    assert _dates.get_dates(fn) == [
+        datetime.datetime(2022, 12, 4, 0, 0, 0),
+        datetime.datetime(2022, 12, 4, 0, 0, 0),
     ]
 
 
 def test_get_dates_with_gdal_string():
     # Checks that is can parse 'NETCDF:"/path/to/file.nc":variable'
     assert _dates.get_dates('NETCDF:"/usr/19990101/20200303_20210101.nc":variable') == [
-        datetime.date(2020, 3, 3),
-        datetime.date(2021, 1, 1),
+        datetime.datetime(2020, 3, 3),
+        datetime.datetime(2021, 1, 1),
     ]
     assert _dates.get_dates(
         'NETCDF:"/usr/19990101/20200303_20210101.nc":"//variable/2"'
     ) == [
-        datetime.date(2020, 3, 3),
-        datetime.date(2021, 1, 1),
+        datetime.datetime(2020, 3, 3),
+        datetime.datetime(2021, 1, 1),
     ]
     # Check the derived dataset name too
     assert _dates.get_dates(
         'DERIVED_SUBDATASET:AMPLITUDE:"/usr/19990101/20200303_20210101.int"'
-    ) == [datetime.date(2020, 3, 3), datetime.date(2021, 1, 1)]
+    ) == [datetime.datetime(2020, 3, 3), datetime.datetime(2021, 1, 1)]
 
 
 def test_sort_files_by_date():
@@ -88,10 +119,10 @@ def test_sort_files_by_date():
         "slc_20180101.tif",
     ]
     expected_dates = [
-        [datetime.date(2018, 1, 1)],
-        [datetime.date(2019, 1, 1)],
-        [datetime.date(2020, 3, 3)],
-        [datetime.date(2021, 1, 1)],
+        [datetime.datetime(2018, 1, 1)],
+        [datetime.datetime(2019, 1, 1)],
+        [datetime.datetime(2020, 3, 3)],
+        [datetime.datetime(2021, 1, 1)],
     ]
     expected_files = sorted(files)
 
@@ -126,10 +157,10 @@ def test_sort_files_by_date_interferograms():
         "ifg_20180101_20210101.tif",
     ]
     dates = [
-        [datetime.date(2018, 1, 1), datetime.date(2021, 1, 1)],
-        [datetime.date(2019, 1, 1), datetime.date(2020, 3, 3)],
-        [datetime.date(2020, 3, 3), datetime.date(2021, 1, 1)],
-        [datetime.date(2020, 3, 3), datetime.date(2022, 1, 1)],
+        [datetime.datetime(2018, 1, 1), datetime.datetime(2021, 1, 1)],
+        [datetime.datetime(2019, 1, 1), datetime.datetime(2020, 3, 3)],
+        [datetime.datetime(2020, 3, 3), datetime.datetime(2021, 1, 1)],
+        [datetime.datetime(2020, 3, 3), datetime.datetime(2022, 1, 1)],
     ]
     sorted_files, sorted_dates = _dates.sort_files_by_date(files)
     assert sorted_files == sorted(files)  # here lexicographic order is correct
@@ -147,12 +178,12 @@ def test_sort_files_by_date_compressed_first():
         "compressed_20200101_20210101.tif",
     ]
     expected_dates = [
-        [datetime.date(2018, 1, 1), datetime.date(2020, 1, 1)],
-        [datetime.date(2020, 1, 1), datetime.date(2021, 1, 1)],
-        [datetime.date(2018, 1, 1)],
-        [datetime.date(2019, 1, 1)],
-        [datetime.date(2020, 1, 1)],
-        [datetime.date(2021, 1, 1)],
+        [datetime.datetime(2018, 1, 1), datetime.datetime(2020, 1, 1)],
+        [datetime.datetime(2020, 1, 1), datetime.datetime(2021, 1, 1)],
+        [datetime.datetime(2018, 1, 1)],
+        [datetime.datetime(2019, 1, 1)],
+        [datetime.datetime(2020, 1, 1)],
+        [datetime.datetime(2021, 1, 1)],
     ]
 
     sorted_files, sorted_dates = _dates.sort_files_by_date(unsorted_files)